+++ packages/visual-grid-client/src/sdk/takeScreenshot.js
@@ -12,21 +12,15 @@ async function takeScreenshot({
   serverUrl,
   proxy,
   renderInfo,
-  cdt,
+  snapshot,
   url,
-  resourceUrls,
-  blobs,
-  frames,
   browsers = [{width: 1024, height: 768}],
   sizeMode = 'full-page',
   selector,
   region,
   scriptHooks,
 }) {
-  const {resourceContents, frames: framesWithResources} = deserializeDomSnapshotResult({
-    blobs,
-    frames,
-  })
+  const snapshots = Array.isArray(snapshot) ? snapshot : Array(browsers.length).fill(snapshot)
 
   const renderingInfo = new RenderingInfo({
     serviceUrl: renderInfo.serviceUrl,
@@ -42,17 +36,26 @@ async function takeScreenshot({
     renderingInfo,
   })
 
-  const {rGridDom: dom, allResources: resources} = await createRGridDOMAndGetResourceMapping({
-    resourceUrls,
-    resourceContents,
-    cdt,
-    frames: framesWithResources,
-  })
+  const results = await Promise.all(
+    snapshots.map(snapshot => {
+      const {resourceUrls, resourceContents, frames, cdt} = deserializeDomSnapshotResult(snapshot)
+      return createRGridDOMAndGetResourceMapping({resourceUrls, resourceContents, frames, cdt})
+    }),
+  )
+
+  const {dom, resources} = results.reduce(
+    (results, {rGridDom, allResources}) => {
+      results.dom.push(rGridDom)
+      results.resources.push(Object.values(allResources))
+      return results
+    },
+    {dom: [], resources: []},
+  )
 
   const renderRequests = createRenderRequests({
     url,
     dom,
-    resources: Object.values(resources),
+    resources,
     browsers,
     renderInfo: renderingInfo,
     sizeMode,