+++ packages/visual-grid-client/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- added support for multiple dom snapshots for `checkWindow`, `testWindow` and `takeScreenshot`
+- removed @applitools/dom-snapshot from dependencies
 - move `takeDomSnapshot` to `eyes-sdk-core`
 
 ## 14.7.5 - 2020/8/13